Update results.csv
Changed Kelly Farrell's deadlift result form 575 to 275.  Clearly there was a notation error on the score-sheet from RPS, most likely in a 2 and 5 being substituted (as they look similar).  This was a "beginners only meet" and for a girl to set a world record deadlift yet only being able squat 250 and bench 120 is highly suspect.  Additionally I have not seen any news about this world record nor have I heard of this person at any other meets so it's very unlikely they are an elite level powerlifter.